requirements.txt
pandas
protobuf
scikit_learn
streamlit
google-generativeai
google-cloud-aiplatform
requests
joblib